Solucionados errores de order

+++ backend/app/model/order.py
@@ -6,8 +6,8 @@ class Order(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer)
     status = db.Column(db.String(20), default="pending")
-    total_amount = db.Column(db.Numeric(10, 2))
-    created_at = db.Column(db.DateTime, default=datetime.utcnow())
+    total_amount = db.Column(db.Float)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
 
 class OrderItem(db.Model):
     __tablename__ = "order_items"
@@ -15,4 +15,4 @@ class OrderItem(db.Model):
     order_id = db.Column(db.Integer)
     product_id = db.Column(db.Integer)
     quantity = db.Column(db.Integer)
-    unit_price = db.Column(db.Numeric(10, 2))
+    unit_price = db.Column(db.Float)
